requirements.txt
torch>=1.9.0
pandas>=1.3.0
numpy>=1.20.0
scikit-learn>=0.24.0
matplotlib>=3.4.0
seaborn>=0.11.0
gym>=0.20.0
tqdm>=4.61.0
pyyaml>=5.4.0
joblib>=1.0.0
xgboost>=1.5.0
scipy>=1.7.0